tests/data/usecases/update-current-account/local-update-current-account.spec.ts
import { SetStorageMock } from '@/tests/data/mocks'
import { mockAccountModel } from '@/tests/domain/mocks'
import { LocalUpdateCurrentAccount } from '@/data/usecases/update-current-account'
import { UnexpectedError } from '@/domain/error'

type SutTypes = {
  sut: LocalUpdateCurrentAccount
  setStorageMock: SetStorageMock
}

const makeSut = (): SutTypes => {
  const setStorageMock = new SetStorageMock()
  const sut = new LocalUpdateCurrentAccount(setStorageMock)
  return {
    sut,
    setStorageMock
  }
}

describe('LocalUpdateCurrentAccount', () => {
  it('Should call SetStorage with correct value', async () => {
    const { sut, setStorageMock } = makeSut()
    const account = mockAccountModel()
    await sut.save(account)
    expect(setStorageMock.key).toBe('account')
    expect(setStorageMock.value).toBe(JSON.stringify(account))
  })

  it('Should throw if SetStorage throws', async () => {
    const { sut, setStorageMock } = makeSut()
    jest.spyOn(setStorageMock, 'set').mockRejectedValueOnce(new Error())
    const promise = sut.save(mockAccountModel())
    expect(promise).rejects.toThrow(new Error())
  })

  it('Should throw if accessToken is falsy', async () => {
    const { sut } = makeSut()
    const promise = sut.save(undefined)
    expect(promise).rejects.toThrow(new UnexpectedError())
  })
})
